Working on new code organization
