+++ src/ipbb/cmds/proj.py
@@ -8,8 +8,9 @@ import subprocess
 
 # Elements
 from ..tools.common import SmartOpen
-from ..defaults import kProjAreaFile, kProjDir, ProjectInfo
-from ..cli.utils import DirSentry, raiseError, validateComponent
+from ..defaults import kProjAreaFile, kProjDir
+from . import ProjectInfo
+from .utils import DirSentry, raiseError, validateComponent
 from ..depparser.Pathmaker import Pathmaker
 
 from os.path import join, split, exists, splitext, relpath, isdir
